feat - Add field on crabe_framework
remove Field structure on the crabe_ssl
