+++ crabe_framework/src/data/field.rs
@@ -1,0 +1,12 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, Copy, Clone)]
+pub struct Field {
+    pub width: f32,
+    pub length: f32,
+    pub goal_width: f32,
+    pub goal_depth: f32,
+    pub penalty_depth: f32,
+    pub penalty_width: f32,
+    pub center_radius: f32,
+}
